Show user/group memories selector on the user/group-related pages
