+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.107.1] - 2022-03-24
+### Added
+- Links to the user/group memories in the sidebar of the user/group-related
+  pages
+
 ### Changed
 - Show registration date in Profiles and Groups
